+++ .gitignore
@@ -1,2 +1,3 @@
 /temp_files
 /__pycache__
+/Logs
